refactor: isolate model eval persistence
