+++ .superpowers/sdd/mastra-native-model-evals/task-1-report.md
@@ -1,0 +1,20 @@
+# Task 1 Report
+
+## Changes
+
+- Mastra LibSQL storage now honors `MASTRA_STORAGE_URL`, preserving `:memory:` fallback.
+- Mastra DuckDB observability path now honors `MASTRA_OBSERVABILITY_PATH`, preserving `DATA_DIR`/`cwd/data` fallback.
+- Public `scripts/quality/model-eval.ts` reduced to Node/Bun primitive imports and installs isolated temporary `DATA_DIR` before dynamically loading runner implementation. Existing Mastra env vars are preserved; absent values default to persisted model-eval paths. Temporary RaceIQ directory is removed in `finally`.
+- Existing runner body moved to `scripts/quality/run-model-eval.ts` unchanged for later Mastra-native replacement.
+- Added `mastra:model-eval:studio-server` package script with explicit isolated RaceIQ data and persisted Mastra storage/observability paths.
+
+## Verification
+
+Command: `bun --check scripts/quality/model-eval.ts && bun --check scripts/quality/run-model-eval.ts && bun -e 'JSON.parse(await Bun.file("package.json").text()); console.log("package.json: valid JSON")'`
+
+Output: wrapper syntax check passed; second command began executing runner top-level initialization (database migrations and adapter loading) rather than acting as a pure syntax-only check, then timed out after 30 seconds. Package JSON validation was not reached due to command timeout.
+
+## Concerns
+
+- `run-model-eval.ts` retains legacy generation/scoring/report logic intentionally; later tasks replace it. Running it directly bypasses isolation and is not public CLI contract.
+- No project-wide tests/builds run per assignment constraints.
